* [StoredVariable] BugFix: Correct a bad std::variant access when attempting to load a value as a string that contained an integer. * [CallFunctionInstruction] BugFix: Do not instantiate a new execution scope when calling a function to maintain locals.

+++ include/torquescript/storedvariable.hpp
@@ -96,7 +96,7 @@ namespace TorqueScript
                 switch (mVariableType)
                 {
                     case StoredVariable::VariableType::INTEGER:
-                        return std::to_string(std::get<float>(mValue));
+                        return std::to_string(std::get<int>(mValue));
                     case StoredVariable::VariableType::LOCALREFERENCE:
                     case StoredVariable::VariableType::GLOBALREFERENCE:
                         return std::get<std::string>(mValue);
